lifecycle_watcher.py initial implmentation
